requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
asyncpgsa==0.27.1
attrs==23.2.0
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
GeoAlchemy2==0.14.4
idna==3.6
iniconfig==2.0.0
multidict==6.0.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-aiohttp==1.0.5
pytest-asyncio==0.23.5
PyYAML==6.0.1
SQLAlchemy==1.4.45
tomli==2.0.1
typing_extensions==4.9.0
yarl==1.9.4
